pyproject.toml
[tool.poetry]
name = "crystalvision"
version = "0.2.0"
description = "FFTCG Image Classification and Detection."
authors = ["Allonte <angelosheltie@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
opencv-contrib-python = ">=4.8.1.78"
keras-nightly = ">=3.1.0.dev2024021703"
keras-tuner = ">=1.4.5"
scikit-learn = ">=1.3.0"
scikit-image = ">=0.22.0"
pandas = ">=2.1.2"
pyarrow = ">=15.0.0"
Pillow = ">=10.1.0"
mlxtend = ">=0.23.0"
tqdm = ">=4.66.1"
aiohttp = ">=3.8.6"
aiofiles = ">=23.2.1"
dogpile-cache = "^1.3.1"


[tool.poetry.group.tensorflow.dependencies]
# tensorflow-io = "0.31.0"
# tensorflow = [
#     {version = "*", platform="linux"},
#     {version = "<=2.11", platform="win32"},
# ]


[tool.poetry.group.torch.dependencies]
torch = {version = "^2.2.0+cu121", source = "pytorch"}
torchvision = {version = "^0.17.0+cu121", source = "pytorch"}
torchaudio = {version = "^2.2.0+cu121", source = "pytorch"}


[tool.poetry.group.dev.dependencies]
pycodestyle = ">=2.11.1"
flake8 = ">=6.1.0"
pydocstyle = ">=6.3.0"
ipykernel = ">=6.26.0"
tensorboard = "<2.11"
pre-commit = "^3.6.0"
black = ">=24.1.1"
ruff = ">=0.1.14"
pytest = ">=8.0.0"
pytest-env = "^1.1.3"



[tool.poetry.group.debug.dependencies]
paretoset = ">=1.2.3"


[tool.poetry.group.labeling.dependencies]
pyqt5-qt5 = [
    {version = "*", platform="linux"},
    {version = "<=5.15.10", platform="win32"},
]
labelme = ">=5.4.1"
pyqt5 = "^5.15.10"
labelme2yolo = ">=0.1.4"


[tool.poetry.group.detection.dependencies]
ultralytics = ">=8.1.11"
onnx = ">=1.12.0"
onnxsim = ">=0.4.33"
onnxruntime-gpu = ">=1.12.0"
lapx = ">=0.5.2"
# ray = {extras = ["tune"], version = ">=2.9.2"}



[tool.poetry.group.accelerate.dependencies]
accimage = {path = "accimage-0.2.0-cp310-cp310-win_amd64.whl"}

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.pytest_env]
KERAS_BACKEND = "torch"

[tool.ruff]
select = ["E", "F"]
ignore = ["E501"]
